Add files via upload
added files 12/15/09:29 before trying to make a bunch of changes that may screw it up. Attempting to get the app to render the number of individual submissions so that it can count how many forms make up the heat map. Will be relacing the app.get for the home.html.